# RadiKeep/Features/Program/ProgramEndpoints.cs
using System.Globalization;
using Microsoft.AspNetCore.Http.HttpResults;
using RadiKeep.Features.Shared.Models;
using RadiKeep.Logics.Context;
using RadiKeep.Logics.Extensions;
using RadiKeep.Logics.Logics.PlayProgramLogic;
using RadiKeep.Logics.Logics.ProgramScheduleLogic;
using RadiKeep.Logics.Logics.RadikoLogic;
using RadiKeep.Logics.Logics.ReserveLogic;
using RadiKeep.Logics.Logics.StationLogic;
using RadiKeep.Logics.Mappers;
using RadiKeep.Logics.Models;
using RadiKeep.Logics.Models.Enums;
using RadiKeep.Logics.Models.NhkRadiru;
using RadiKeep.Logics.Primitives.DataAnnotations;
using RadiKeep.Logics.Services;
using ZLogger;

namespace RadiKeep.Features.Program;

/// <summary>
/// 番組関連の Api エンドポイントを提供する。
/// </summary>
public static class ProgramEndpoints
{
    /// <summary>
    /// 番組関連エンドポイントをマッピングする。
    /// </summary>
    public static IEndpointRouteBuilder MapProgramEndpoints(this IEndpointRouteBuilder endpoints)
    {
        var group = endpoints.MapGroup("/api/programs").WithTags("ProgramsApi");
        group.MapGet("/stations/radiko", HandleGetRadikoStationsAsync)
            .WithName("ApiProgramStationsRadiko")
            .WithSummary("radiko放送局一覧を取得する");
        group.MapGet("/stations/radiru", HandleGetRadiruStationsAsync)
            .WithName("ApiProgramStationsRadiru")
            .WithSummary("らじる放送局一覧を取得する");
        group.MapGet("/list/radiko", HandleGetRadikoProgramsAsync)
            .WithName("ApiProgramListRadiko")
            .WithSummary("radiko番組一覧を取得する");
        group.MapGet("/list/radiru", HandleGetRadiruProgramsAsync)
            .WithName("ApiProgramListRadiru")
            .WithSummary("らじる番組一覧を取得する");
        group.MapGet("/now", HandleGetNowOnAirProgramsAsync)
            .WithName("ApiProgramNow")
            .WithSummary("現在放送中の番組を取得する");
        group.MapGet("/detail", HandleGetProgramDetailAsync)
            .WithName("ApiProgramDetail")
            .WithSummary("番組詳細を取得する");
        group.MapPost("/search", HandleSearchProgramsAsync)
            .WithName("ApiProgramSearch")
            .WithSummary("番組を検索する");
        group.MapPost("/keyword-reserve", HandleSetKeywordReserveAsync)
            .WithName("ApiProgramKeywordReserve")
            .WithSummary("キーワード予約を登録する");
        group.MapPost("/update", HandleUpdateProgramsAsync)
            .WithName("ApiProgramUpdate")
            .WithSummary("番組表更新ジョブを起動する");
        group.MapPost("/reserve", HandleReserveProgramAsync)
            .WithName("ApiProgramReserve")
            .WithSummary("番組録音予約を登録する");
        group.MapPost("/play", HandlePlayProgramAsync)
            .WithName("ApiProgramPlay")
            .WithSummary("番組再生情報を取得する");
        return endpoints;
    }

    /// <summary>
    /// 現在エリアのradiko放送局ID一覧を取得する。
    /// </summary>
    private static async ValueTask<List<string>> GetCurrentAreaStationsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        RadikoUniqueProcessLogic radikoUniqueProcessLogic,
        StationLobLogic stationLobLogic)
    {
        var (isSuccess, area) = await radikoUniqueProcessLogic.GetRadikoAreaAsync();
        if (!isSuccess || string.IsNullOrWhiteSpace(area))
        {
            logger.ZLogWarning($"radikoエリア情報の取得に失敗");
            return [];
        }

        return await stationLobLogic.GetCurrentAreaStations(area);
    }

    /// <summary>
    /// radiko放送局一覧を取得する。
    /// </summary>
    private static async Task<Ok<ApiResponse<Dictionary<string, List<RadikoStationInformationEntry>>>>> HandleGetRadikoStationsAsync(
        StationLobLogic stationLobLogic)
    {
        var stations = (await stationLobLogic.GetRadikoStationAsync())
            .GroupBy(r => r.RegionId)
            .OrderBy(r => r.First().RegionOrder)
            .ToDictionary(r => r.First().RegionName, r => r.OrderBy(s => s.StationOrder).ToList());
        return TypedResults.Ok(ApiResponse.Ok(stations));
    }

    /// <summary>
    /// らじる放送局一覧を取得する。
    /// </summary>
    private static Ok<ApiResponse<Dictionary<string, List<RadiruStationEntry>>>> HandleGetRadiruStationsAsync(
        StationLobLogic stationLobLogic)
    {
        var stations = stationLobLogic.GetRadiruStationAsync()
            .GroupBy(r => r.AreaName)
            .ToDictionary(group => group.Key, group => group.ToList());
        return TypedResults.Ok(ApiResponse.Ok(stations));
    }

    /// <summary>
    /// radiko番組一覧を取得する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<List<RadioProgramEntry>>>, BadRequest<ApiResponse<EmptyData?>>>> HandleGetRadikoProgramsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        IAppConfigurationService config,
        RadikoUniqueProcessLogic radikoUniqueProcessLogic,
        StationLobLogic stationLobLogic,
        ProgramScheduleLobLogic programScheduleLobLogic,
        string d,
        string s)
    {
        if (!DateOnly.TryParseExact(d, "yyyyMMdd", null, DateTimeStyles.None, out var date))
        {
            logger.ZLogWarning($"日付の変換に失敗 {d}");
            return TypedResults.BadRequest(ApiResponse.Fail("日付の変換に失敗しました。"));
        }

        if (!config.IsRadikoAreaFree)
        {
            var currentAreaStations = await GetCurrentAreaStationsAsync(logger, radikoUniqueProcessLogic, stationLobLogic);
            var currentAreaStationSet = currentAreaStations.ToHashSet(StringComparer.OrdinalIgnoreCase);
            if (currentAreaStationSet.Count == 0 || !currentAreaStationSet.Contains(s))
            {
                return TypedResults.Ok(ApiResponse.Ok(new List<RadioProgramEntry>()));
            }
        }

        var programs = await programScheduleLobLogic.GetRadikoProgramListAsync(date, s);
        return TypedResults.Ok(ApiResponse.Ok(programs));
    }

    /// <summary>
    /// らじる番組一覧を取得する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<List<RadioProgramEntry>>>, BadRequest<ApiResponse<EmptyData?>>>> HandleGetRadiruProgramsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        ProgramScheduleLobLogic programScheduleLobLogic,
        string d,
        string s,
        string a)
    {
        if (!DateOnly.TryParseExact(d, "yyyyMMdd", null, DateTimeStyles.None, out var date))
        {
            logger.ZLogWarning($"日付の変換に失敗 {d}");
            return TypedResults.BadRequest(ApiResponse.Fail("日付の変換に失敗しました。"));
        }

        var programs = await programScheduleLobLogic.GetRadiruProgramAsync(date, a, s);
        return TypedResults.Ok(ApiResponse.Ok(programs));
    }

    /// <summary>
    /// 現在放送中の番組一覧を取得する。
    /// </summary>
    private static async Task<Ok<ApiResponse<ProgramNowOnAirResponse>>> HandleGetNowOnAirProgramsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        IRadioAppContext appContext,
        IAppConfigurationService config,
        StationLobLogic stationLobLogic,
        RadikoUniqueProcessLogic radikoUniqueProcessLogic,
        ProgramScheduleLobLogic programScheduleLobLogic)
    {
        var radikoPrograms = await programScheduleLobLogic.GetRadikoNowOnAirProgramListAsync(appContext.StandardDateTimeOffset);
        var radiruPrograms = await programScheduleLobLogic.GetRadiruNowOnAirProgramListAsync(appContext.StandardDateTimeOffset);

        var programs = radikoPrograms.Concat(radiruPrograms).OrderBy(r => r.StationName).ToList();
        List<string> currentAreaStationsForUi = [];

        var stationList = await stationLobLogic.GetAllRadikoStationAsync();
        var stationById = stationList.ToDictionary(r => r.StationId, r => r);
        var regionOrderMap = stationList
            .GroupBy(r => r.RegionId)
            .ToDictionary(r => r.Key, r => new { RegionName = r.First().RegionName, RegionOrder = r.Min(s => s.RegionOrder) });

        foreach (var program in programs)
        {
            if (stationById.TryGetValue(program.StationId, out var station))
            {
                program.AreaId = station.RegionId;
                program.AreaName = station.RegionName;
            }
        }

        currentAreaStationsForUi = await GetCurrentAreaStationsAsync(logger, radikoUniqueProcessLogic, stationLobLogic);
        if (!config.IsRadikoAreaFree)
        {
            var currentAreaStationSet = currentAreaStationsForUi.ToHashSet(StringComparer.OrdinalIgnoreCase);
            programs = programs
                .Where(p => p.ServiceKind == RadioServiceKind.Radiru || currentAreaStationSet.Contains(p.StationId))
                .ToList();
        }

        var radiruAreaOrderMap = Enum
            .GetValues<RadiruAreaKind>()
            .Select((area, index) => new { AreaId = area.GetEnumCodeId(), AreaOrder = index })
            .ToDictionary(x => x.AreaId, x => x.AreaOrder);

        var areaMap = new Dictionary<string, ProgramAreaEntry>(StringComparer.OrdinalIgnoreCase);
        foreach (var program in programs)
        {
            if (string.IsNullOrWhiteSpace(program.AreaId) || string.IsNullOrWhiteSpace(program.AreaName))
            {
                continue;
            }

            var key = $"{(int)program.ServiceKind}:{program.AreaId}";
            if (areaMap.ContainsKey(key))
            {
                continue;
            }

            if (program.ServiceKind == RadioServiceKind.Radiko && regionOrderMap.TryGetValue(program.AreaId, out var region))
            {
                areaMap[key] = new ProgramAreaEntry(program.AreaId, region.RegionName, region.RegionOrder, 0);
                continue;
            }

            if (program.ServiceKind == RadioServiceKind.Radiru && radiruAreaOrderMap.TryGetValue(program.AreaId, out var radiruOrder))
            {
                areaMap[key] = new ProgramAreaEntry(program.AreaId, program.AreaName, radiruOrder, 1);
                continue;
            }

            areaMap[key] = new ProgramAreaEntry(program.AreaId, program.AreaName, int.MaxValue, 9);
        }

        var areaList = areaMap.Values
            .OrderBy(r => r.ServiceOrder)
            .ThenBy(r => r.AreaOrder)
            .ThenBy(r => r.AreaName, StringComparer.OrdinalIgnoreCase)
            .ToList();

        var data = new ProgramNowOnAirResponse(
            programs,
            areaList,
            currentAreaStationsForUi,
            config.IsRadikoAreaFree);
        return TypedResults.Ok(ApiResponse.Ok(data));
    }

    /// <summary>
    /// 番組詳細を取得する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<RadioProgramEntry?>>, BadRequest<ApiResponse<EmptyData?>>>> HandleGetProgramDetailAsync(
        ProgramScheduleLobLogic programScheduleLobLogic,
        string? id,
        string? kind)
    {
        if (id == null || kind == null)
        {
            return TypedResults.BadRequest(ApiResponse.Fail("必要な項目が記載されていません。"));
        }

        var radioServiceKind = kind.GetEnumByCodeId<RadioServiceKind>();
        if (radioServiceKind == RadioServiceKind.Undefined)
        {
            return TypedResults.BadRequest(ApiResponse.Fail("サービス種別が不正です。"));
        }

        var program = await programScheduleLobLogic.GetProgramAsync(id, radioServiceKind);
        return TypedResults.Ok(ApiResponse.Ok(program));
    }

    /// <summary>
    /// 番組検索を実行する。
    /// </summary>
    private static async Task<Ok<ApiResponse<IEnumerable<ProgramForApiEntry>>>> HandleSearchProgramsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        IAppConfigurationService config,
        RadikoUniqueProcessLogic radikoUniqueProcessLogic,
        StationLobLogic stationLobLogic,
        ProgramScheduleLobLogic programScheduleLobLogic,
        IEntryMapper entryMapper,
        ProgramSearchEntity entity)
    {
        var result = new List<ProgramForApiEntry>();

        if (entity.SelectedRadikoStationIds.Any())
        {
            if (!config.IsRadikoAreaFree)
            {
                var currentAreaStations = await GetCurrentAreaStationsAsync(logger, radikoUniqueProcessLogic, stationLobLogic);
                var currentAreaStationSet = currentAreaStations.ToHashSet(StringComparer.OrdinalIgnoreCase);
                entity.SelectedRadikoStationIds = entity.SelectedRadikoStationIds
                    .Where(id => currentAreaStationSet.Contains(id))
                    .ToList();
            }

            if (entity.SelectedRadikoStationIds.Any())
            {
                var list = await programScheduleLobLogic.SearchRadikoProgramAsync(entity);
                result.AddRange(list.Select(entryMapper.ToRadikoProgramForApiEntry));
            }
        }

        if (entity.SelectedRadiruStationIds.Any())
        {
            var list = await programScheduleLobLogic.SearchRadiruProgramAsync(entity);
            result.AddRange(list.Select(entryMapper.ToRadiruProgramForApiEntry));
        }

        result = entity.OrderKind switch
        {
            KeywordReserveOrderKind.ProgramStartDateTimeAsc => result.OrderBy(x => x.StartTime).ToList(),
            KeywordReserveOrderKind.ProgramStartDateTimeDesc => result.OrderByDescending(x => x.StartTime).ToList(),
            KeywordReserveOrderKind.ProgramEndDateTimeAsc => result.OrderBy(x => x.EndTime).ToList(),
            KeywordReserveOrderKind.ProgramEndDateTimeDesc => result.OrderByDescending(x => x.EndTime).ToList(),
            KeywordReserveOrderKind.ProgramNameAsc => result.OrderBy(x => x.Title).ToList(),
            KeywordReserveOrderKind.ProgramNameDesc => result.OrderByDescending(x => x.Title).ToList(),
            _ => result
        };

        return TypedResults.Ok(ApiResponse.Ok(result.Take(100)));
    }

    /// <summary>
    /// キーワード予約を登録する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<EmptyData?>>, BadRequest<ApiResponse<EmptyData?>>>> HandleSetKeywordReserveAsync(
        ReserveLobLogic reserveLobLogic,
        KeywordReserveEntry entry)
    {
        if (!Enum.IsDefined(typeof(KeywordReserveTagMergeBehavior), entry.MergeTagBehavior))
        {
            return TypedResults.BadRequest(ApiResponse.Fail("タグマージ設定が不正です。"));
        }

        if (!string.IsNullOrEmpty(entry.RecordPath) && !entry.RecordPath.IsValidRelativePath())
        {
            return TypedResults.BadRequest(ApiResponse.Fail("保存先パスが不正です。"));
        }

        if (!string.IsNullOrEmpty(entry.RecordFileName) && !entry.RecordFileName.IsValidFileName())
        {
            return TypedResults.BadRequest(ApiResponse.Fail("ファイル名が不正です。"));
        }

        var (isSuccess, error) = await reserveLobLogic.SetKeywordReserveAsync(entry);
        if (!isSuccess)
        {
            return TypedResults.BadRequest(ApiResponse.Fail(error?.Message ?? "キーワード予約の登録に失敗しました。"));
        }

        return TypedResults.Ok(ApiResponse.Ok("登録しました。"));
    }

    /// <summary>
    /// 番組表更新ジョブを起動する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<EmptyData?>>, BadRequest<ApiResponse<EmptyData?>>>> HandleUpdateProgramsAsync(
        ILogger<ProgramEndpointsMarker> logger,
        IServiceScopeFactory serviceScopeFactory)
    {
        try
        {
            // API は即時応答し、番組表更新本体は独立スコープでバックグラウンド実行する。
            // スコープを作り直して、リクエスト終了後の破棄済み DbContext 参照を防ぐ。
            _ = Task.Run(
                async () =>
                {
                    try
                    {
                        using var scope = serviceScopeFactory.CreateScope();
                        var programUpdateRunner = scope.ServiceProvider.GetRequiredService<ProgramUpdateRunner>();
                        await programUpdateRunner.ExecuteAsync("manual");
                    }
                    catch (Exception ex)
                    {
                        logger.ZLogError(ex, $"番組表更新バックグラウンド実行でエラーが発生しました。");
                    }
                });
        }
        catch (Exception ex)
        {
            logger.ZLogError(ex, $"番組表更新処理の起動に失敗しました。");
            return TypedResults.BadRequest(ApiResponse.Fail("番組表更新ジョブの起動に失敗しました。"));
        }

        return TypedResults.Ok(ApiResponse.Ok("番組表更新処理を実行中です。通常数分で更新が完了します。"));
    }

    /// <summary>
    /// 番組の録音予約を登録する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<EmptyData?>>, BadRequest<ApiResponse<EmptyData?>>>> HandleReserveProgramAsync(
        ReserveLobLogic reserveLobLogic,
        ProgramInformationRequestEntry program)
    {
        (bool isSuccess, Exception? error) result = program.RadioServiceKind switch
        {
            RadioServiceKind.Radiko => await reserveLobLogic.SetRecordingJobByProgramIdAsync(program.ProgramId, RadioServiceKind.Radiko, program.RecordingType),
            RadioServiceKind.Radiru => await reserveLobLogic.SetRecordingJobByProgramIdAsync(program.ProgramId, RadioServiceKind.Radiru, program.RecordingType),
            _ => (false, new InvalidOperationException($"サービス種別が不正です。 {program.RadioServiceKind}"))
        };

        if (!result.isSuccess)
        {
            var message = program.RadioServiceKind is RadioServiceKind.Other or RadioServiceKind.Undefined
                ? "サービス種別が不正です。"
                : result.error?.Message ?? "録音予約に失敗しました。";
            return TypedResults.BadRequest(ApiResponse.Fail(message));
        }

        return TypedResults.Ok(ApiResponse.Ok("予約しました。"));
    }

    /// <summary>
    /// 番組再生情報を取得する。
    /// </summary>
    private static async Task<Results<Ok<ApiResponse<ProgramPlaybackInfoResponse>>, BadRequest<ApiResponse<EmptyData?>>>> HandlePlayProgramAsync(
        PlayProgramLobLogic playProgramLobLogic,
        ProgramInformationRequestEntry program)
    {
        switch (program.RadioServiceKind)
        {
            case RadioServiceKind.Radiko:
            {
                var (isSuccess, token, url, error) = await playProgramLobLogic.PlayRadikoProgramAsync(program.ProgramId);
                if (!isSuccess)
                {
                    return TypedResults.BadRequest(ApiResponse.Fail(error?.Message ?? "再生準備に失敗しました。"));
                }

                return TypedResults.Ok(ApiResponse.Ok(new ProgramPlaybackInfoResponse(token, url)));
            }
            case RadioServiceKind.Radiru:
            {
                var (isSuccess, token, url, error) = await playProgramLobLogic.PlayRadiruProgramAsync(program.ProgramId);
                if (!isSuccess)
                {
                    return TypedResults.BadRequest(ApiResponse.Fail(error?.Message ?? "番組の再生ができませんでした。"));
                }

                return TypedResults.Ok(ApiResponse.Ok(new ProgramPlaybackInfoResponse(token, url)));
            }
            default:
                return TypedResults.BadRequest(ApiResponse.Fail("サービス種別が不正です。"));
        }
    }

    /// <summary>
    /// ProgramEndpoints 用のロガーカテゴリ型。
    /// </summary>
    private sealed class ProgramEndpointsMarker;

    /// <summary>
    /// エリア一覧表示用の内部モデル。
    /// </summary>
    private sealed record ProgramAreaEntry(string AreaId, string AreaName, int AreaOrder, int ServiceOrder);

    /// <summary>
    /// 現在放送中番組一覧レスポンス。
    /// </summary>
    private sealed record ProgramNowOnAirResponse(
        List<RadioProgramEntry> Programs,
        List<ProgramAreaEntry> Areas,
        List<string> CurrentAreaStations,
        bool IsAreaFree);

    /// <summary>
    /// 再生開始に必要な情報レスポンス。
    /// </summary>
    private sealed record ProgramPlaybackInfoResponse(string? Token, string? Url);
}