# RadiKeep/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RadiKeep | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1:7262/"
    }
  ],
  "paths": {
...
    "/api/programs/list/radiko": {
      "get": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "radiko番組一覧を取得する",
        "operationId": "ApiProgramListRadiko",
        "parameters": [
          {
            "name": "d",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfRadioProgramEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/list/radiru": {
      "get": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "らじる番組一覧を取得する",
        "operationId": "ApiProgramListRadiru",
        "parameters": [
          {
            "name": "d",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "s",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfRadioProgramEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/programs/detail": {
      "get": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "番組詳細を取得する",
        "operationId": "ApiProgramDetail",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRadioProgramEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/search": {
      "post": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "番組を検索する",
        "operationId": "ApiProgramSearch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramSearchEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfIEnumerableOfProgramForApiEntry"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/keyword-reserve": {
      "post": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "キーワード予約を登録する",
        "operationId": "ApiProgramKeywordReserve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordReserveEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/update": {
      "post": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "番組表更新ジョブを起動する",
        "operationId": "ApiProgramUpdate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/reserve": {
      "post": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "番組録音予約を登録する",
        "operationId": "ApiProgramReserve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramInformationRequestEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/programs/play": {
      "post": {
        "tags": [
          "ProgramsApi"
        ],
        "summary": "番組再生情報を取得する",
        "operationId": "ApiProgramPlay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramInformationRequestEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfProgramPlaybackInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfEmptyData": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmptyData"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfIEnumerableOfProgramForApiEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ProgramForApiEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfListOfRadioProgramEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RadioProgramEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfRadioProgramEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RadioProgramEntry"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DaysOfWeek": {
        "type": "integer"
      },
...
      "EmptyData": {
        "type": "object"
      },
...
      "KeywordReserveEntry": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Ulid"
          },
          "selectedRadikoStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "selectedRadiruStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keyword": {
            "type": "string"
          },
          "excludedKeyword": {
            "type": "string"
          },
          "recordPath": {
            "type": "string"
          },
          "recordFileName": {
            "type": "string"
          },
          "searchTitleOnly": {
            "type": "boolean"
          },
          "excludeTitleOnly": {
            "type": "boolean"
          },
          "selectedDaysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DaysOfWeek"
            }
          },
          "startTimeString": {
            "type": "string"
          },
          "endTimeString": {
            "type": "string"
          },
          "startTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "endTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "startDelay": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "endDelay": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mergeTagBehavior": {
            "$ref": "#/components/schemas/KeywordReserveTagMergeBehavior"
          }
        }
      },
...
      "KeywordReserveTagMergeBehavior": {
        "type": "integer"
      },
...
      "ProgramInformationRequestEntry": {
        "required": [
          "programId"
        ],
        "type": "object",
        "properties": {
          "programId": {
            "type": "string"
          },
          "radioServiceKind": {
            "$ref": "#/components/schemas/RadioServiceKind"
          },
          "recordingType": {
            "$ref": "#/components/schemas/RecordingType"
          }
        }
      },
...
      "ProgramSearchEntity": {
        "type": "object",
        "properties": {
          "serviceKind": {
            "type": "string"
          },
          "selectedRadikoStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "selectedRadiruStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keyword": {
            "type": "string"
          },
          "excludedKeyword": {
            "type": "string"
          },
          "searchTitleOnly": {
            "type": "boolean"
          },
          "searchTitleOnlyExcludedKeyword": {
            "type": "boolean"
          },
          "selectedDaysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DaysOfWeek"
            }
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "includeHistoricalPrograms": {
            "type": "boolean"
          },
          "orderKind": {
            "type": "string"
          }
        }
      },
...
      "Ulid": { },
...
    }
  },
...
}